Add files via upload
'Project Netflix 101' is an interactive project whereby I analyzed our Netflix history, focusing on primarily two things:

which days of the week had a particular show been watched a great deal, and which days not;
what time of the day had a particular show been watched a great deal, and when not. This has been achieved through the analysis of the data and the creation of two distinct graphs.